+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.3 [unreleased]
 
+- Added support for missing values
 - Fixed Daru training and prediction
 
 ## 0.1.2